requirements.txt
fastapi
uvicorn
streamlit
google-generativeai
python-dotenv
python-multipart
pydantic